PromptEngineering/pyproject.toml
[tool.poetry]
name = "PromptEngineering"
version = "0.2.0"
description = "Ejemplo uso programatico AI"
authors = ["Mario Codes <mariocodes1993@gmail.com>"]
readme = "README.md"
packages = [
	{ include = "CallToOpenAI.py" },
	{ include = "Embeddings" },
	{ include = "LangChain" },
]

[tool.poetry.dependencies]
python = "^3.14"
python-dotenv = "^1.2.1"
dotenv = "^0.9.9"
pandas = "^2.3.3"
openai = "^2.11.0"
langchain-openai = "^1.1.7"
langchain = "^1.2.3"
azure-cosmos = "^4.15.0"

[tool.poetry.scripts]
call-to-openai = "CallToOpenAI:main"
call-to-get-embeddings = "Embeddings.ManualCallToGetEmbeddingsFromText:main"
call-to-insert-embeddings = "Embeddings.CallToInsertEmbeddingsIntoCosmos:main"
call-to-query-embeddings = "Embeddings.CallToQueryEmbeddingsFromCosmos:main"

langchain = "LangChain.LangChain:main"
langchain-lcel = "LangChain.LangChainExpressionLanguage:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
